+++ assets/block/src/block.json
@@ -1,0 +1,17 @@
+{
+	"$schema": "https://schemas.wp.org/trunk/block.json",
+	"apiVersion": 2,
+	"name": "wp-contributions/block",
+	"version": "0.1.0",
+	"title": "WP Contributions Block",
+	"category": "common",
+	"icon": "embed-photo",
+	"description": "Wordpress Contributions Showcase made easy.",
+	"supports": {
+		"html": true
+	},
+	"textdomain": "wp-contributions-block",
+	"editorScript": "file:./index.js",
+	"editorStyle": "file:./index.css",
+	"style": "file:./style-index.css"
+}
