+++ assets/block/src/index.js
@@ -1,6 +1,52 @@
 /**
- * Imports
+ * Registers a new block provided a unique name and an object defining its behavior.
+ *
+ * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-registration/
+ */
+import { registerBlockType } from '@wordpress/blocks';
+/**
+ * Lets webpack process CSS, SASS or SCSS files referenced in JavaScript files.
+ * All files containing `style` keyword are bundled together. The code used
+ * gets applied both to the front of your site and to the editor.
+ *
+ * @see https://www.npmjs.com/package/@wordpress/scripts#using-css
  */
 import './style.scss';
-import './block.js';
-import './sidebar.js';
+
+/**
+ * Internal dependencies
+ */
+import Edit from './edit';
+import save from './save';
+import metadata from './block.json';
+
+/**
+ * Every block starts by registering a new block type definition.
+ *
+ * @see https://developer.wordpress.org/block-editor/reference-guides/block-api/block-registration/
+ */
+registerBlockType( metadata.name, {
+	attributes: {
+		slug: {
+			type: 'string'
+		},
+		preferred_username: {
+			type: 'string',
+		},
+		theme: {
+			type: 'boolean',
+			default: false,
+		},
+		contribution_type: {
+			type: 'string',
+			default: 'plugin',
+		}
+	},
+	/**
+	 * @see ./edit.js
+	 */
+	edit: Edit,
+	/**
+	 * @see ./save.js
+	 */
+} );
