+++ assets/block/src/style.scss
@@ -1,144 +1,12 @@
-.click-to-send {
-	display: block;
-	background-color: #fff;
-	position: relative;
-	border: 1px solid #dddddd;
-	-moz-border-radius: 4px;
-	border-radius: 4px;
-	padding: 15px 30px;
-	margin: 15px 0px;
+/**
+ * The following styles get applied both on the front of your site
+ * and in the editor.
+ *
+ * Replace them with your own styles or remove the file completely.
+ */
 
-	.ctt-type {
-		width: 30%;
-		height: 70px;
-		color: #999999;
-	}
-
-	.ctt-textbox {
-		border-bottom: 1px solid #333;
-		padding-bottom: 10px;
-		margin-top: 30px;
-	}
-
-	.ctt-text {
-		margin: 0 0 10px 0;
-		padding: 0;
-		position: relative;
-		word-wrap: break-word;
-		color: #999999;
-		font-size: 14px;
-		line-height: 140%;
-		box-shadow: none;
-		letter-spacing: 0.05em;
-		font-weight: 100;
-		text-decoration: none;
-		text-transform: none;
-
-		a {
-			margin: 0;
-			padding: 0;
-			position: relative;
-			color: #999999;
-			font-size: 24px;
-			line-height: 140%;
-			box-shadow: none;
-			letter-spacing: 0.05em;
-			font-weight: 100;
-			text-decoration: none;
-			text-transform: none;
-		}
-
-		select {
-			line-height: normal;
-		}
-
-	}
-
-	a.ctt-btn {
-		margin: 0;
-		padding: 0;
-		padding-right: 0px;
-		position: relative;
-		display: block;
-		text-transform: uppercase;
-		font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-		box-shadow: none;
-		font-size: 12px;
-		font-weight: 100;
-		line-height: 100%;
-		color: #999999;
-		float: right;
-		text-decoration: none;
-
-	}
-}
-
-.click-to-send-alt {
-	display: block;
-	background-color: #333333;
-	position: relative;
-	border: 1px solid #333333;
-	-moz-border-radius: 4px;
-	border-radius: 4px;
-	padding: 15px 30px;
-	margin: 15px 0px;
-
-	.ctt-text {
-		margin: 0 0 10px 0;
-		padding: 0;
-		position: relative;
-		word-wrap: break-word;
-		color: #FFF;
-		font-size: 14px;
-		line-height: 140%;
-		box-shadow: none;
-		letter-spacing: 0.05em;
-		font-weight: 100;
-		text-decoration: none;
-		text-transform: none;
-
-		a {
-			margin: 0;
-			padding: 0;
-			position: relative;
-			color: #FFF;
-			font-size: 14px;
-			line-height: 140%;
-			box-shadow: none;
-			letter-spacing: 0.05em;
-			font-weight: 100;
-			text-decoration: none;
-			text-transform: none;
-
-			&:hover {
-				color: #EEEEEE;
-			}
-		}
-	}
-
-	a.ctt-btn {
-		margin: 0;
-		padding: 0;
-		padding-right: 0px;
-		position: relative;
-		display: block;
-		text-transform:uppercase;
-		font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-		box-shadow: none;
-		font-size: 12px;
-		font-weight: 100;
-		line-height: 100%;
-		color: #FFF;
-		float: right;
-		text-decoration: none;
-
-		&:hover {
-			color: #EEEEEE;
-		}
-
-	}
-}
-
-.css-1opxiwx.css-1opxiwx.css-1opxiwx {
-	line-height: normal !important;
+.wp-block-create-block-contrib-block {
+	background-color: #21759b;
+	color: #fff;
+	padding: 2px;
 }
